api/requirements.txt
Django==4.0.1
psycopg2==2.9.3
djangorestframework==3.13.1
django-cors-headers==3.11.0